separate out cohort analysis models from cohort results models
