+++ src/main/java/org/ohdsi/webapi/cohortanalysis/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Model specific classes used by for running Heracles analyses (cohort analyses)
+ */
+package org.ohdsi.webapi.cohortanalysis;
