+++ src/main/java/org/ohdsi/webapi/service/CohortAnalysisService.java
@@ -18,10 +18,10 @@ import jersey.repackaged.com.google.common.base.Joiner;
 import org.ohdsi.sql.SqlRender;
 import org.ohdsi.sql.SqlSplit;
 import org.ohdsi.sql.SqlTranslate;
-import org.ohdsi.webapi.cohortresults.CohortAnalysis;
-import org.ohdsi.webapi.cohortresults.CohortAnalysisTask;
-import org.ohdsi.webapi.cohortresults.CohortAnalysisTasklet;
-import org.ohdsi.webapi.cohortresults.CohortSummary;
+import org.ohdsi.webapi.cohortanalysis.CohortAnalysis;
+import org.ohdsi.webapi.cohortanalysis.CohortAnalysisTask;
+import org.ohdsi.webapi.cohortanalysis.CohortAnalysisTasklet;
+import org.ohdsi.webapi.cohortanalysis.CohortSummary;
 import org.ohdsi.webapi.helper.ResourceHelper;
 import org.ohdsi.webapi.job.JobExecutionResource;
 import org.ohdsi.webapi.job.JobTemplate;
